Move CLI logic to a separate module
